requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3

